+++ service/controller/key/provider.go
@@ -1,5 +1,11 @@
 package key
 
+import "strings"
+
 func IsAWS(provider string) bool {
 	return provider == "aws"
+}
+
+func IsAWSChina(region string) bool {
+	return strings.HasPrefix(region, "cn-")
 }