+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Create `external-dns-cluster-values` configmap on cluster creation.
+
 ## [5.8.0] - 2023-09-01
 
 ### Added