+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.8.0-patch1] - 2023-09-26
+
 ### Added
 
 - Create `external-dns-cluster-values` configmap on cluster creation.
@@ -477,7 +479,8 @@ be edited by users.
 
 
 
-[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v5.8.0...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-operator/compare/v5.8.0-patch1...HEAD
+[5.8.0-patch1]: https://github.com/giantswarm/cluster-operator/compare/v5.8.0...v5.8.0-patch1
 [5.8.0]: https://github.com/giantswarm/cluster-operator/compare/v5.7.1...v5.8.0
 [5.7.1]: https://github.com/giantswarm/cluster-operator/compare/v5.7.0...v5.7.1
 [5.7.0]: https://github.com/giantswarm/cluster-operator/compare/v5.6.1...v5.7.0